pyproject.toml
[project]
name = "repl-nix-robotremotecontrol"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-socketio>=5.5.0",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
]
